+++ src/pages/Users/components/UserRow/style.js
@@ -1,0 +1,11 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useRowStyles = makeStyles({
+  root: {
+    '& > *': {
+      borderBottom: 'unset',
+    },
+  },
+});
+
+export default useRowStyles;
